requirements.txt
requests
python-dotenv
colorama
wikipedia-api